Revert "VisitDaoImplTest implemented, subtle errors in DAO classes corrected"
This reverts commit 93c22532a5cc8c27f71360f187a25014b5249f9b.

+++ MushroomHunterPortal/src/main/java/cz/fi/muni/pa165/mushroomhunter/dao/VisitDao.java
@@ -24,7 +24,7 @@ public interface VisitDao {
 
     public void delete(Visit visit);
 
-    public Visit find(long id);
+    public Visit find(int id);
 
     public List<Visit> findByHunter(Hunter hunter);
